[Code] Start of sprite renderer + consolidate all inspectors into single file
