+++ Projects/Editor/Resources/Editor.prefs
@@ -1,4 +1,4 @@
-Scene View: '00: 0.97510105 01: -3.7252903e-09 02: -0.2217611 03: 0 10: 0.07166445 11: 0.94634414 12: 0.3151142 13: 0 20: 0.20986232 21: -0.3231606 22: 0.92278117 23: 0 30: -0.57498825 31: 3.276701 32: -21.958 33: 1'
+Scene View: '00: 0.9979762 01: 0 02: -0.06358886 03: 0 10: -0.003120158 11: 0.99879545 12: -0.04896838 13: 0 20: 0.063512266 21: 0.04906768 22: 0.9967741 23: 0 30: -0.6808611 31: 2.669111 32: -2.8496249 33: 1'
 Startup Project: 'C:/Git/Odyssey/Projects/Sandbox'
 Editor Registry: Resources/AssetRegistry.osettings
 Folder Icon: 213981209753892768
