--- Projects/Editor/Source/GUI/Inspectors/Inspector.h
@@ -1,13 +1,0 @@
-#pragma once
-#include "Ref.h"
-#include "EditorEnums.h"
-#include "GUIElement.h"
-
-namespace Odyssey
-{
-	class Inspector : public GUIElement
-	{
-	public:
-		virtual bool Draw() { return false; }
-	};
-}
