--- Projects/Editor/Source/GUI/Inspectors/MeshInspector.h
@@ -1,30 +1,0 @@
-#pragma once
-#include "Inspector.h"
-#include "Mesh.h"
-#include "PropertyDrawers.h"
-
-namespace Odyssey
-{
-	class MeshInspector : public Inspector
-	{
-	public:
-		MeshInspector() = default;
-		MeshInspector(GUID guid);
-
-	public:
-		virtual bool Draw() override;
-
-	private:
-		void OnNameChanged(std::string_view name);
-		void OnSourceAssetChanged(GUID sourceGUID);
-
-	private:
-		Ref<Mesh> m_Mesh;
-		StringDrawer m_GUIDDrawer;
-		StringDrawer m_NameDrawer;
-		StringDrawer m_TypeDrawer;
-		StringDrawer m_VertexCountDrawer;
-		StringDrawer m_IndexCountDrawer;
-		AssetFieldDrawer m_SourceMeshDrawer;
-	};
-}
