+++ Projects/Editor/imgui.ini
@@ -1,5 +1,5 @@
 [Window][Debug##Default]
-Pos=1541,255
+Pos=255,205
 Size=1247,875
 Collapsed=0
 
@@ -55,7 +55,7 @@ DockId=0x00000019,1
 Pos=405,0
 Size=1131,1080
 Collapsed=0
-DockId=0x0000001A,0
+DockId=0x00000014,0
 
 [Window][Scene Hierarchy Window]
 Size=403,1080
@@ -71,7 +71,7 @@ Collapsed=0
 Pos=405,0
 Size=1131,798
 Collapsed=0
-DockId=0x0000001A,0
+DockId=0x00000014,0
 
 [Window][Content Browser]
 Pos=517,878
@@ -87,9 +87,9 @@ DockId=0x0000000E,0
 
 [Window][Scene View]
 Pos=517,75
-Size=949,801
+Size=2182,801
 Collapsed=0
-DockId=0x0000001A,0
+DockId=0x00000014,0
 
 [Window][Scene Hierarchy]
 Pos=0,75
@@ -101,7 +101,7 @@ DockId=0x00000015,0
 Pos=354,19
 Size=1134,688
 Collapsed=0
-DockId=0x0000001A,1
+DockId=0x00000014,1
 
 [Window][Dear ImGui Metrics/Debugger]
 Pos=1929,96
@@ -109,10 +109,10 @@ Size=1175,1185
 Collapsed=0
 
 [Window][Game View]
-Pos=1468,75
-Size=1231,801
+Pos=517,75
+Size=2182,801
 Collapsed=0
-DockId=0x0000001B,0
+DockId=0x00000014,1
 
 [Window][Example: Documents]
 ViewportPos=988,281
@@ -185,7 +185,7 @@ Collapsed=0
 Pos=1667,73
 Size=1211,801
 Collapsed=0
-DockId=0x0000001A,1
+DockId=0x00000014,1
 
 [Window][Scene Hierarchy(2)]
 Pos=936,486
@@ -234,7 +234,7 @@ Collapsed=0
 Pos=517,75
 Size=949,801
 Collapsed=0
-DockId=0x0000001A,1
+DockId=0x00000014,1
 
 [Window][Example: Auto-resizing window]
 Pos=204,218
@@ -323,33 +323,31 @@ Column 1  Width=132
 Column 2  Width=198
 
 [Docking][Data]
-DockNode                    ID=0x00000019 Pos=449,249 Size=1358,802 Selected=0x766639F3
-DockSpace                   ID=0x8B93E3BD Window=0xA787BDB4 Pos=0,59 Size=3440,1321 Split=Y
-  DockNode                  ID=0x0000000F Parent=0x8B93E3BD SizeRef=1920,43 Selected=0xE6059F3F
-  DockNode                  ID=0x00000010 Parent=0x8B93E3BD SizeRef=1920,1002 Split=Y
-    DockNode                ID=0x00000005 Parent=0x00000010 SizeRef=1920,94 Selected=0xE3DFA36A
-    DockNode                ID=0x00000006 Parent=0x00000010 SizeRef=1920,951 Split=X
-      DockNode              ID=0x00000007 Parent=0x00000006 SizeRef=515,1047 Split=Y Selected=0x413E6147
-        DockNode            ID=0x00000015 Parent=0x00000007 SizeRef=459,802 Selected=0x2E9237F7
-        DockNode            ID=0x00000016 Parent=0x00000007 SizeRef=459,474 Selected=0x413E6147
-      DockNode              ID=0x00000008 Parent=0x00000006 SizeRef=2923,1047 Split=X
-        DockNode            ID=0x0000000D Parent=0x00000008 SizeRef=2182,1047 Split=X
-          DockNode          ID=0x00000003 Parent=0x0000000D SizeRef=1536,1080 Split=X
-            DockNode        ID=0x00000001 Parent=0x00000003 SizeRef=403,1080 Selected=0x4EEFFB03
-            DockNode        ID=0x00000002 Parent=0x00000003 SizeRef=1131,1080 Split=Y Selected=0x7E1E4FEF
-              DockNode      ID=0x00000009 Parent=0x00000002 SizeRef=1887,801 Split=X Selected=0x08AA43A5
-                DockNode    ID=0x00000013 Parent=0x00000009 SizeRef=399,776 Selected=0xE6E7F153
-                DockNode    ID=0x00000014 Parent=0x00000009 SizeRef=754,776 Split=X Selected=0x08AA43A5
-                  DockNode  ID=0x0000001A Parent=0x00000014 SizeRef=949,801 CentralNode=1 Selected=0x08AA43A5
-                  DockNode  ID=0x0000001B Parent=0x00000014 SizeRef=1231,801 Selected=0x642CEEBB
-              DockNode      ID=0x0000000A Parent=0x00000002 SizeRef=1887,473 Split=X Selected=0xBF096F38
-                DockNode    ID=0x00000011 Parent=0x0000000A SizeRef=1287,474 Selected=0xBF096F38
-                DockNode    ID=0x00000012 Parent=0x0000000A SizeRef=1142,474 Selected=0xA2F76E8A
-          DockNode          ID=0x00000004 Parent=0x0000000D SizeRef=382,1080 Selected=0xF58E28BB
-        DockNode            ID=0x0000000E Parent=0x00000008 SizeRef=739,1047 Selected=0xE7039252
-DockSpace                   ID=0xE5EFEDAE Pos=517,164 Size=949,741 Split=X
-  DockNode                  ID=0x0000000B Parent=0xE5EFEDAE SizeRef=475,746 Split=Y Selected=0xC9A8EE47
-    DockNode                ID=0x00000017 Parent=0x0000000B SizeRef=396,305 Selected=0x14312729
-    DockNode                ID=0x00000018 Parent=0x0000000B SizeRef=396,439 Selected=0xC9A8EE47
-  DockNode                  ID=0x0000000C Parent=0xE5EFEDAE SizeRef=820,746 CentralNode=1 Selected=0xE3F3CE3C
+DockNode                  ID=0x00000019 Pos=449,249 Size=1358,802 Selected=0x766639F3
+DockSpace                 ID=0x8B93E3BD Window=0xA787BDB4 Pos=0,59 Size=3440,1321 Split=Y
+  DockNode                ID=0x0000000F Parent=0x8B93E3BD SizeRef=1920,43 Selected=0xE6059F3F
+  DockNode                ID=0x00000010 Parent=0x8B93E3BD SizeRef=1920,1002 Split=Y
+    DockNode              ID=0x00000005 Parent=0x00000010 SizeRef=1920,94 Selected=0xE3DFA36A
+    DockNode              ID=0x00000006 Parent=0x00000010 SizeRef=1920,951 Split=X
+      DockNode            ID=0x00000007 Parent=0x00000006 SizeRef=515,1047 Split=Y Selected=0x413E6147
+        DockNode          ID=0x00000015 Parent=0x00000007 SizeRef=459,802 Selected=0x2E9237F7
+        DockNode          ID=0x00000016 Parent=0x00000007 SizeRef=459,474 Selected=0x413E6147
+      DockNode            ID=0x00000008 Parent=0x00000006 SizeRef=2923,1047 Split=X
+        DockNode          ID=0x0000000D Parent=0x00000008 SizeRef=2182,1047 Split=X
+          DockNode        ID=0x00000003 Parent=0x0000000D SizeRef=1536,1080 Split=X
+            DockNode      ID=0x00000001 Parent=0x00000003 SizeRef=403,1080 Selected=0x4EEFFB03
+            DockNode      ID=0x00000002 Parent=0x00000003 SizeRef=1131,1080 Split=Y Selected=0x7E1E4FEF
+              DockNode    ID=0x00000009 Parent=0x00000002 SizeRef=1887,801 Split=X Selected=0x08AA43A5
+                DockNode  ID=0x00000013 Parent=0x00000009 SizeRef=399,776 Selected=0xE6E7F153
+                DockNode  ID=0x00000014 Parent=0x00000009 SizeRef=754,776 CentralNode=1 Selected=0x08AA43A5
+              DockNode    ID=0x0000000A Parent=0x00000002 SizeRef=1887,473 Split=X Selected=0xBF096F38
+                DockNode  ID=0x00000011 Parent=0x0000000A SizeRef=1287,474 Selected=0xBF096F38
+                DockNode  ID=0x00000012 Parent=0x0000000A SizeRef=1142,474 Selected=0xA2F76E8A
+          DockNode        ID=0x00000004 Parent=0x0000000D SizeRef=382,1080 Selected=0xF58E28BB
+        DockNode          ID=0x0000000E Parent=0x00000008 SizeRef=739,1047 Selected=0xE7039252
+DockSpace                 ID=0xE5EFEDAE Pos=517,164 Size=949,741 Split=X
+  DockNode                ID=0x0000000B Parent=0xE5EFEDAE SizeRef=475,746 Split=Y Selected=0xC9A8EE47
+    DockNode              ID=0x00000017 Parent=0x0000000B SizeRef=396,305 Selected=0x14312729
+    DockNode              ID=0x00000018 Parent=0x0000000B SizeRef=396,439 Selected=0xC9A8EE47
+  DockNode                ID=0x0000000C Parent=0xE5EFEDAE SizeRef=820,746 CentralNode=1 Selected=0xE3F3CE3C
